requirements.txt
asgiref==3.10.0
black==25.11.0
click==8.3.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==23.0.0
mypy_extensions==1.1.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.3.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
celery==5.4.0
redis==5.2.1
channels
channels-redis
daphne
Pillow==11.0.0
reportlab==4.2.5
xhtml2pdf==0.2.16

